requirements.txt
cffi==1.14.0
click==7.1.1
cryptography==2.9.2
Cython==0.29.16
eventlet==0.25.2
Flask==1.1.2
Flask-Cors==3.0.8
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-engineio==3.12.1
python-socketio==4.5.1
PyYAML==5.3.1
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
WTForms==2.3.1
